fix: refacto torrent lib
