+++ streaming_service/src/torrentMetaParser.ts
@@ -2,20 +2,10 @@ import * as fs from "fs";
 import https from "https";
 import bencode from "bencode";
 import * as ttypes from "./ft_torrent_types.js";
-import { createHash, randomBytes } from "node:crypto";
 
-function generateInfoHash(torrentPath: string): string {
-  const torrent = fs.readFileSync(torrentPath);
-  const toHash = bencode.decode(torrent);
-
-  return encodeURIComponent(
-    createHash("sha1").update(bencode.encode(toHash)).digest("hex"),
-  );
-}
-
-function generatePeerId(): string {
-  const peerId = "-FT1000-" + randomBytes(12).toString("hex");
-  return encodeURIComponent(createHash("sha1").update(peerId).digest("hex"));
+function generatePath(torrentUrl: string): string {
+  const splitedUrl: string[] = torrentUrl.split("/");
+  return `./torrents/${splitedUrl[splitedUrl.length - 1].toLowerCase()}.torrent`;
 }
 
 function filterNonHtppTrackers(trackers: string[][]): string[][] {
@@ -31,33 +21,6 @@ function filterNonHtppTrackers(trackers: string[][]): string[][] {
   }
 
   return httpTrackers;
-}
-
-export async function queryTracker(
-  torrentPath: string,
-  torrentMetaData: ttypes.TorrentMeta,
-) {
-  const infoHash = generateInfoHash(torrentPath);
-  const peerId = generatePeerId();
-  const ports: number[] = [];
-
-  for (let i = 6881; i < 6889; i++) {
-    ports.push(i);
-  }
-
-  const uploaded = 0;
-  const downloaded = 0;
-  const left = torrentMetaData.info.pieces.length;
-  const query = `${torrentMetaData.announce}?peer_id=${peerId}&info_hash=${infoHash}&port=${ports[0]}&uploaded=${uploaded}&downloaded=${downloaded}&left=${left}`;
-
-  console.log(`Querying: ${query}`);
-  const response = await fetch(query);
-  return response.text();
-}
-
-function generatePath(torrentUrl: string): string {
-  const splitedUrl: string[] = torrentUrl.split("/");
-  return `./torrents/${splitedUrl[splitedUrl.length - 1].toLowerCase()}.torrent`;
 }
 
 export function downloadTorrent(torrentUrl: string): string {
